% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Construct an object of type Dataset}
\usage{
Dataset(file, path)
}
\arguments{
\item{file}{An object of type File which hosts the dataset}

\item{path}{The complete intrafile path to the dataset}
}
\description{
A Dataset is a representation of a dataset in a HDF5 file.
}
\examples{
src <- Source('http://hsdshdflab.hdfgroup.org')
f <- File(src, '/home/spollack/testzero.h5')
d <- Dataset(f, '/grpA/grpAB/dsetX')
}
