% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\docType{methods}
\name{dimnames,RHDF5ArraySeed-method}
\alias{dimnames,RHDF5ArraySeed-method}
\title{HDF server content is assumed transposed relative to R matrix layout.}
\usage{
\S4method{dimnames}{RHDF5ArraySeed}(x)
}
\description{
HDF server content is assumed transposed relative to R matrix layout.
}
