% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{domainContents}
\alias{domainContents}
\title{The h5serv API does not have a GET /domains method, so we need
to descend the tree to the requested domain node by UUID's.}
\usage{
domainContents(object, rootdir = "/hdfgroup/org")
}
\description{
The h5serv API does not have a GET /domains method, so we need
to descend the tree to the requested domain node by UUID's.
}
