% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\docType{methods}
\name{domains}
\alias{domains}
\alias{domains,Source,character-method}
\alias{domains,Source,missing-method}
\title{List files and subdirectories of a domain}
\usage{
domains(object, rootdir)

\S4method{domains}{Source,character}(object, rootdir)

\S4method{domains}{Source,missing}(object)
}
\arguments{
\item{object}{An object of type Source}

\item{rootdir}{A slash-separated directory in the Source file system.}
}
\value{
a vector of domains in the rootdir
}
\description{
The user needs to give the domain to start in. The search
will be non-recursive. I.e., output for domain '/home/jreadey/' will 
not return the files in '/home/jreadey/HDFLabTutorial/'
}
\examples{
src.hsds <- Source('http://hsdshdflab.hdfgroup.org')
src.chan <- Source('http://h5s.channingremotedata.org:5000', 'h5serv')
domains(src.chan)
domains(src.hsds, '/home/jreadey')
}
