% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,Dataset-method}
\title{extract elements of a one or two-dimensional Dataset}
\usage{
\S4method{[}{Dataset,numeric,ANY}(x, i)
}
\description{
extract elements of a one or two-dimensional Dataset
}
