% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\docType{class}
\name{File-class}
\alias{File-class}
\title{An S4 class to represent an HDF5 file accessible from a server.}
\description{
An S4 class to represent an HDF5 file accessible from a server.
}
\section{Slots}{

\describe{
\item{\code{Source}}{an object of type Source}

\item{\code{domain}}{the file's domain on the server; more or less, an alias for its 
location in the external server file system}

\item{\code{dsetdf}}{a data.frame that caches often-used information about the file}
}}

