% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{File}
\alias{File}
\title{Construct an object of type File}
\usage{
File(src, domain)
}
\arguments{
\item{src}{an object of type Source, the server which exposes the file}

\item{domain}{the domain string; the file's location on the server's
file system.}
}
\description{
A File is a representation of an HDF5 file the contents of which are accessible 
exposed by a HDF5 server.
}
\examples{
src <- Source('http://hsdshdflab.hdfgroup.org')
f10x <- File(src, '/shared/bioconductor/tenx_full.h5')
}
