% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{findDatasets}
\alias{findDatasets}
\title{private - traverse internal file hiearchy, find datasets, and
cache often-accessed information in a data.frame for the File object.}
\usage{
findDatasets(src, domain)
}
\description{
private - traverse internal file hiearchy, find datasets, and
cache often-accessed information in a data.frame for the File object.
}
