% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,Dataset,character,character-method}
\alias{getData,Dataset,character,missing-method}
\alias{getData,Dataset,list,character-method}
\alias{getData,Dataset,list,missing-method}
\alias{getData,Dataset,list,,missing-method}
\title{Fetch data from a remote dataset}
\usage{
getData(dataset, indices, transfermode)

\S4method{getData}{Dataset,character,character}(dataset, indices,
  transfermode)

\S4method{getData}{Dataset,character,missing}(dataset, indices)

\S4method{getData}{Dataset,list,character}(dataset, indices, transfermode)

\S4method{getData}{Dataset,list,missing}(dataset, indices)
}
\arguments{
\item{dataset}{An object of type Dataset, the dataset to access.}

\item{indices}{The indices of the data to fetch}

\item{transfermode}{Either (default) 'JSON' or 'binary'}
}
\value{
an Array containing the data fetched from the server
}
\description{
The servers require data to be fetched in slices, i.e., in sets of 
for which the indices of each dimension are of the form start:stop:step.
More complex sets of indices will be split into slices and fetched in
multiple requests. This is opaque to the user, but may enter into 
considerations of data access patterns, e.g., for performance-tuning.
}
\examples{
s <- Source('http://hsdshdflab.hdfgroup.org')
f <- File(s, '/shared/bioconductor/tenx_full.h5')
d <- Dataset(f, '/newassay001')
x <- getData(d, c('1:4', '1:27998'), transfermode='JSON')
# x <- getData(d, c(1:4, 1:27998), transfermode='JSON') # method missing?
x <- d[1:4,1:27998]
}
