% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\docType{class}
\name{RHDF5ArraySeed}
\alias{RHDF5ArraySeed}
\alias{RHDF5ArraySeed-class}
\title{RHDF5ArraySeed for RHDF5Array backend to DelayedArray}
\usage{
RHDF5ArraySeed(endpoint, svrtype, domain, dsetname)
}
\arguments{
\item{endpoint}{URL of remote server}

\item{svrtype}{type of server, must be either 'hsds' or 'h5serv'}

\item{domain}{HDF5 domain of H5 file on server}

\item{dsetname}{complete internal path to dataset in H5 file}
}
\value{
An initialized object of type RHDF5ArraySeed
}
\description{
RHDF5ArraySeed for RHDF5Array backend to DelayedArray

Construct an object of type RHDF5ArraySeed
}
\section{Slots}{

\describe{
\item{\code{endpoint}}{URL of remote server}

\item{\code{svrtype}}{type of server, must be either 'hsds' or 'h5serv'}

\item{\code{domain}}{HDF5 domain of H5 file on server}

\item{\code{dsetname}}{complete internal path to dataset in H5 file}

\item{\code{dataset}}{object of type Dataset for access to the H5 dataset}
}}

