% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{Source}
\alias{Source}
\title{Construct an object of type Source.}
\usage{
Source(endpoint, type = "hsds")
}
\arguments{
\item{endpoint}{URL for server}

\item{type}{Type of server software at the source; must be}
}
\value{
An object of type Source
}
\description{
A Source is a representation of a URL which provides access to a HDF5 
server (either h5serv or hsds.)
}
\examples{
src.hsds <- Source('http://hsdshdflab.hdfgroup.org')
src.chan <- Source('http://h5s.channingremotedata.org:5000', 'h5serv')
}
