% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,HSDSDataset-method}
\alias{[,HSDSDataset,numeric-method}
\alias{[,HSDSDataset,numeric,numeric-method}
\alias{[,HSDSDataset,numeric,ANY-method}
\alias{[,HSDSDataset,numeric,ANY,ANY-method}
\title{extract elements of a one or two-dimensional HSDSDataset}
\usage{
\S4method{[}{HSDSDataset,numeric,ANY}(x, i)
}
\arguments{
\item{x}{object of type HSDSDataset}

\item{i}{vector of indices (first dimension)}
}
\description{
extract elements of a one or two-dimensional HSDSDataset
}
